requirements.txt
djangorestframework
django-cors-headers
pymysql
